refactor(atomic-a11y): decompose reporter and openacr modules, convert scripts to TypeScript
Tasks 2, 3, 5 completed:

**Task 2: Decompose vitest-a11y-reporter.ts**
- Created src/reporter/reporter-utils.ts (337 lines)
- Created src/reporter/axe-integration.ts (58 lines)
- Modified vitest-a11y-reporter.ts to import from new modules

**Task 3: Decompose json-to-openacr.ts**
- Created src/openacr/types.ts (195 lines)
- Created src/openacr/overrides.ts (79 lines)
- Created src/openacr/manual-audit.ts (208 lines)
- Created src/openacr/conformance.ts (135 lines)
- Created src/openacr/report-builder.ts (289 lines)
- Created src/openacr/yaml-serializer.ts (103 lines)
- Reduced json-to-openacr.ts to 137 lines ✓ Under 200 target

**Task 5: Convert scripts to TypeScript**
- Converted 3 scripts from .mjs to .ts
- All scripts use shared imports

TypeScript compiles successfully
