+++ packages/atomic-a11y/src/openacr/conformance.ts
@@ -1,0 +1,135 @@
+import type {A11yCriterionReport} from '../shared/types.js';
+import {resolveManualConformance} from './manual-audit.js';
+import type {
+  A11yOverrideEntry,
+  CriterionAggregate,
+  ManualAuditAggregate,
+  OpenAcrConformance,
+} from './types.js';
+import {reportConformanceToOpenAcr} from './types.js';
+
+const DEFAULT_MANUAL_PLACEHOLDER_NOTE =
+  'Manual audit pending. Phase 3 results will be merged into this report.';
+
+export function mapCriterionConformance(
+  criterion: A11yCriterionReport | undefined
+): OpenAcrConformance | null {
+  if (!criterion) {
+    return null;
+  }
+
+  return reportConformanceToOpenAcr[criterion.conformance] ?? null;
+}
+
+function resolveAutomatedConformance(
+  aggregate: CriterionAggregate | undefined
+): OpenAcrConformance {
+  const coveredCount = aggregate?.coveredComponents.size ?? 0;
+  if (coveredCount === 0) {
+    return 'not-evaluated';
+  }
+
+  const violatingCount = aggregate?.violatingComponents.size ?? 0;
+  if (violatingCount === 0) {
+    return 'supports';
+  }
+
+  if (violatingCount >= coveredCount) {
+    return 'does-not-support';
+  }
+
+  return 'partially-supports';
+}
+
+export function resolveConformance(
+  criterion: A11yCriterionReport | undefined,
+  aggregate: CriterionAggregate | undefined,
+  manualAggregate: ManualAuditAggregate[] | undefined,
+  override: A11yOverrideEntry | undefined
+): OpenAcrConformance {
+  if (override) {
+    return override.conformance;
+  }
+
+  const manualConformance = resolveManualConformance(manualAggregate);
+  if (manualConformance) {
+    return manualConformance;
+  }
+
+  const existingConformance = mapCriterionConformance(criterion);
+
+  if (existingConformance && existingConformance !== 'not-evaluated') {
+    return existingConformance;
+  }
+
+  return resolveAutomatedConformance(aggregate);
+}
+
+export function buildRemarks(
+  criterionId: string,
+  conformance: OpenAcrConformance,
+  coveredComponents: string[],
+  violatingComponents: string[],
+  manualAggregates: ManualAuditAggregate[] | undefined,
+  override: A11yOverrideEntry | undefined
+): string {
+  if (override) {
+    return `[Override] ${override.reason}`;
+  }
+
+  if (manualAggregates && manualAggregates.length > 0) {
+    let passCount = 0;
+    let failCount = 0;
+    let partialCount = 0;
+    let notApplicableCount = 0;
+
+    for (const aggregate of manualAggregates) {
+      switch (aggregate.conformance) {
+        case 'supports':
+          passCount++;
+          break;
+        case 'does-not-support':
+          failCount++;
+          break;
+        case 'partially-supports':
+          partialCount++;
+          break;
+        case 'not-applicable':
+          notApplicableCount++;
+          break;
+      }
+    }
+
+    const componentsCount = manualAggregates.length;
+    const summaryParts: string[] = [];
+    if (passCount > 0) summaryParts.push(`${passCount} pass`);
+    if (failCount > 0) summaryParts.push(`${failCount} fail`);
+    if (partialCount > 0) summaryParts.push(`${partialCount} partial`);
+    if (notApplicableCount > 0)
+      summaryParts.push(`${notApplicableCount} not-applicable`);
+
+    const summary = summaryParts.join(', ');
+    return `Manual audit: ${summary} across ${componentsCount} component(s).`;
+  }
+
+  const coveredCount = coveredComponents.length;
+  const violatingCount = violatingComponents.length;
+
+  if (conformance === 'supports') {
+    return `Automated testing found no axe-core violations for WCAG ${criterionId} across ${coveredCount} mapped component(s). ${DEFAULT_MANUAL_PLACEHOLDER_NOTE}`;
+  }
+
+  if (conformance === 'partially-supports') {
+    return `Automated testing found violations for WCAG ${criterionId} in ${violatingCount} of ${coveredCount} mapped component(s). ${DEFAULT_MANUAL_PLACEHOLDER_NOTE}`;
+  }
+
+  if (conformance === 'does-not-support') {
+    return `Automated testing found violations for WCAG ${criterionId} in all ${coveredCount} mapped component(s). ${DEFAULT_MANUAL_PLACEHOLDER_NOTE}`;
+  }
+
+  if (conformance === 'not-applicable') {
+    return `WCAG ${criterionId} is not applicable for the tested component scope.`;
+  }
+
+  return `WCAG ${criterionId} has no automated mapping evidence in the JSON report. ${DEFAULT_MANUAL_PLACEHOLDER_NOTE}`;
+}
