+++ packages/atomic/.sisyphus/boulder.json
@@ -1,0 +1,9 @@
+{
+  "active_plan": "/Users/ylakhdar/Repos/ui-kit/packages/atomic/.sisyphus/plans/pr7116-decomposition.md",
+  "started_at": "2026-02-14T03:34:36.118Z",
+  "session_ids": [
+    "ses_3a5c90f3bffeVksG4VODDvZ4pY"
+  ],
+  "plan_name": "pr7116-decomposition",
+  "agent": "atlas"
+}
