+++ packages/atomic/.sisyphus/notepads/pr7116-decomposition/decisions.md
@@ -1,0 +1,5 @@
+# Decisions — PR 7116 Decomposition
+
+Architectural choices and rationale.
+
+---
