+++ packages/atomic/.sisyphus/notepads/pr7116-decomposition/issues.md
@@ -1,0 +1,5 @@
+# Issues — PR 7116 Decomposition
+
+Problems, gotchas, and edge cases encountered.
+
+---
