+++ packages/atomic/.sisyphus/notepads/pr7116-decomposition/problems.md
@@ -1,0 +1,5 @@
+# Problems — PR 7116 Decomposition
+
+Unresolved blockers and open questions.
+
+---
